pyproject.toml
[project]
name = "tts-vc"
version = "0.0.1"
description = "TTS and Voice Cloning with Chirp 3 and Cosyvoice"
readme = "README.md"
requires-python = ">=3.12.11"
license = {file = "LICENSE"}
authors = [
    {name = "tts-vc", email = "mghifary@gmail.com"}
]
dependencies = [
    "numpy>=2.0.2",
    "librosa==0.11.0",
    "torch==2.8.0",
    "torchaudio==2.8.0",
    "transformers==4.57.0",
    "google-cloud-texttospeech",
    "gradio==5.49.0",
]

[project.urls]
Homepage = "https://github.com/ghif/tts-vc"
Repository = "https://github.com/ghif/tts-vc"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
